+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 - Improve Data Source to check Organization (#6)
 - Add Notifications and increase Test Coverage (#7)
 - Add Annotation filters (#8)
+- Add Annotation Alert States filter (#9)